project/requirements.txt
matplotlib==3.6.0
moviepy==1.0.3
numpy==1.23.3
Pillow==9.2.
torch~=1.12.1
torchvision~=0.13.1
opencv-python==4.6.0.66
pandas==1.5.0
ipython==8.5.0
pyyaml==6.0
seaborn==0.12.1
